+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased changes
+* [Issue 1148](https://github.com/AmpersandTarski/Ampersand/issues/1148) Allow exec-engine delpair function to handle multiple src and/or tgt atoms
 
 ## v1.8.6 (7 feb 2021)
 * Update EasyRdf library to v1.x